+++ src/JetFormat/Jet3Format.ts
@@ -1,7 +1,9 @@
 import { GENERAL_97_SORT_ORDER } from "../SortOrder";
-import { JetFormat } from "./types";
+import { CodecType, JetFormat } from "./types";
 
 export const jet3Format: JetFormat = {
+    codecType: CodecType.JET,
+
     pageSize: 2048,
 
     textEncoding: "utf8",