+++ src/PageDecrypter/IdentityDecrypter.ts
@@ -1,0 +1,5 @@
+import { PageDecrypter } from "./types";
+
+export function createIdentityDecrypter(): PageDecrypter {
+    return (b) => b;
+}